tests/submission/PhilipTaylor/canvas/2d.text.font.parse.system.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.text.font.parse.system</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.font.html">font</a>.<a href="index.2d.text.font.parse.html">parse</a>.system</h1>
<p class="desc">System fonts must be computed to explicit values</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.text.font.parse">2d.text.font.parse</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.text.font.get">2d.text.font.get</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.text.font.systemfonts">2d.text.font.systemfonts</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("System fonts must be computed to explicit values");
_addTest(function(canvas, ctx) {

ctx.font = 'message-box';
_assertDifferent(ctx.font, 'message-box', "ctx.font", "'message-box'");


});
</script>

<div id="log"></div>
</body></html>
